requirements.txt
astroid==2.11.7
black==22.6.0
click==8.1.3
dill==0.3.5.1
Flask==2.2.1
importlib-metadata==4.12.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pylint==2.14.5
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
Werkzeug==2.2.1
wrapt==1.14.1
zipp==3.8.1

oauthlib~=3.2.1
python-dotenv~=0.21.0